# Template entropy coding functions to help optimiser