This is part of a series of commits authored by @maj160 to improve performance of rav1e.

This commit results in an overall performance improvement of about 2% at speed 2 10-bit:

```
Benchmark 1: ~/Downloads/rav1e_master -s 2 --quantizer 64 ~/xiph-media-files/objective-1-fast-10bit/speed_bag_640x360_60f.y4m -y -o /dev/null
  Time (mean ± σ):     43.704 s ±  0.148 s    [User: 43.702 s, System: 0.183 s]
  Range (min … max):   43.480 s … 43.933 s    10 runs

Benchmark 2: ~/Downloads/rav1e_mod -s 2 --quantizer 64 ~/xiph-media-files/objective-1-fast-10bit/speed_bag_640x360_60f.y4m -y -o /dev/null
  Time (mean ± σ):     42.789 s ±  0.067 s    [User: 42.806 s, System: 0.170 s]
  Range (min … max):   42.679 s … 42.856 s    10 runs
```